+++ npbench/benchmarks/polybench/adi/adi_parir.py
@@ -55,9 +55,9 @@ def kernel(TSTEPS, N, u):
     mul1 = B1 * DT / (DX * DX)
     mul2 = B2 * DT / (DY * DY)
 
-    a = c = torch.tensor(-mul1 / 2.0, dtype=torch.float64)
-    b = e = torch.tensor(1.0 + mul2, dtype=torch.float64)
-    d = f = torch.tensor(-mul2 / 2.0, dtype=torch.float64)
+    a = c = torch.tensor(-mul1 / 2.0, dtype=u.dtype)
+    b = e = torch.tensor(1.0 + mul2, dtype=u.dtype)
+    d = f = torch.tensor(-mul2 / 2.0, dtype=u.dtype)
     par = { 'i': parir.threads(N-2) }
     parir_kernel(u, v, p, q, a, b, c, d, e, f, TSTEPS, N, opts=parir.parallelize(par))
     return u
